Fix the timeout algorithm on i386

+++ retry.go
@@ -102,7 +102,7 @@ func Backoff(operation func() (*Response, error), options ...Option) error {
 		// See the following article...
 		// http://www.awsarchitectureblog.com/2015/03/backoff.html
 		temp := math.Min(capLevel, base*math.Exp2(float64(attempt)))
-		sleepDuration := time.Duration(int(temp/2) + rand.Intn(int(temp/2)))
+		sleepDuration := time.Duration(int64(temp/2) + rand.Intn(int64(temp/2)))
 
 		if sleepDuration < opts.waitTime {
 			sleepDuration = opts.waitTime
